third_party/WebKit/LayoutTests/http/tests/devtools/sources/pretty-print-html-3-expected.txt

Verifies JavaScript pretty-printing functionality.


Running: testLinkFollowedByComment
====== 8< ------
<link href='a/b/c.css' rel='stylesheet'>
<!-- some comment -->

------ >8 ======
Correct mapping for <stylesheet>
Correct mapping for <some>
Correct mapping for <comment>

Running: testInlineJavascript
====== 8< ------
<html>
    <script type="text/javascript">
        for (var i = 0; i < 10; ++i)
            console.log('test ' + i);
    </script>
</html>

------ >8 ======
Correct mapping for <console>
Correct mapping for <test>
Correct mapping for <</html>

Running: testInlineCSS
====== 8< ------
<html>
    <style>
        div {
            color: red;
            border: 1px solid black;
        }
    </style>
</html>

------ >8 ======
Correct mapping for <<html>>
Correct mapping for <red>
Correct mapping for <black>

Running: testMultilineInput
====== 8< ------
<html>
    <head>
        <meta name="ROBOTS" content="NOODP">
        <meta name='viewport' content='text/html'>
        <title>foobar</title>
    <body>
        <script>
            if (1 < 2) {
                if (2 < 3) {
                    if (3 < 4) {
                        if (4 < 5) {
                            console.log("magic")
                        }
                    }
                }
            }
        </script>
        <img src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADIA...">
        <style>
            div {
                display: flex;
                align-items: center;
                justify-content: center;
            }

            body {
                width: 100%
            }

            * {
                border: 1px solid black
            }
        </style>
    </body>
</html>

------ >8 ======
Correct mapping for <ROBOTS>
Correct mapping for <image>
Correct mapping for <...>
Correct mapping for <</body>>
Correct mapping for <</html>>
Correct mapping for <</style>>
